+++ yukon-client/common/src/com/cannontech/common/validation/dao/RphTagDao.java
@@ -6,6 +6,6 @@ import com.cannontech.common.validation.model.RphTag;
 
 public interface RphTagDao {
 
-    boolean insertTag(int changeId, RphTag value);
+    boolean insertTag(long changeId, RphTag value);
     public int clearTagsAfter(long lastChangeIdProcessed, Set<RphTag> tags);
 }
